+++ src/main/java/org/folio/circulation/domain/RequestFulfilmentPreference.java
@@ -1,0 +1,7 @@
+package org.folio.circulation.domain;
+
+public class RequestFulfilmentPreference {
+  private RequestFulfilmentPreference() { }
+
+  public static final String HOLD_SHELF = "Hold Shelf";
+}