+++ src/main/java/org/folio/circulation/support/CommonFailures.java
@@ -4,8 +4,16 @@ import io.vertx.ext.web.RoutingContext;
 import org.folio.circulation.support.http.server.JsonResponse;
 import org.folio.circulation.support.http.server.ServerErrorResponse;
 
+import java.util.function.Consumer;
+
 public class CommonFailures {
   private CommonFailures() { }
+
+  public static Consumer<String> reportFailureToClient(
+    RoutingContext routingContext) {
+
+    return f -> ServerErrorResponse.internalError(routingContext.response(), f);
+  }
 
   public static void reportFailureToFetchInventoryRecords(
     RoutingContext routingContext,