+++ src/test/java/org/folio/circulation/api/requests/scenarios/SingleOpenHoldShelfRequestTests.java
@@ -1,0 +1,39 @@
+package org.folio.circulation.api.requests.scenarios;
+
+import org.folio.circulation.api.support.APITests;
+import org.folio.circulation.support.http.client.IndividualResource;
+import org.folio.circulation.support.http.client.Response;
+import org.junit.Test;
+
+import java.net.MalformedURLException;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeoutException;
+
+import static org.folio.circulation.api.support.builders.RequestBuilder.OPEN_AWAITING_PICKUP;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.junit.MatcherAssert.assertThat;
+
+public class SingleOpenHoldShelfRequestTests extends APITests {
+  @Test
+  public void statusChangesToAwaitingPickupWhenItemCheckedIn()
+    throws InterruptedException,
+    MalformedURLException,
+    TimeoutException,
+    ExecutionException {
+
+    IndividualResource smallAngryPlanet = itemsFixture.basedUponSmallAngryPlanet();
+    IndividualResource james = usersFixture.james();
+    IndividualResource jessica = usersFixture.jessica();
+
+    IndividualResource loanToJames = loansFixture.checkOut(smallAngryPlanet, james);
+
+    IndividualResource requestByJessica = requestsFixture.placeHoldShelfRequest(
+      smallAngryPlanet, jessica);
+
+    loansFixture.checkIn(loanToJames);
+
+    Response request = requestsClient.getById(requestByJessica.getId());
+
+    assertThat(request.getJson().getString("status"), is(OPEN_AWAITING_PICKUP));
+  }
+}